requirements.txt
fastapi>=0.110,<1.0
uvicorn>=0.23,<1.0
pydantic>=2.7,<3
httpx>=0.27,<0.28
PyYAML>=6.0,<7
pytest>=8,<9
hypothesis>=6.100,<7
schemathesis>=3.24,<4
ruff>=0.4,<1
build>=1.2,<2
prometheus-client>=0.20,<1
python-json-logger>=2.0,<3
